+++ src/components/BsField/BsNumericField.vue
@@ -160,7 +160,7 @@
 <script>
 import { FontAwesomeIcon } from "@fortawesome/vue-fontawesome";
 import BsIcon from "../BsIcon/BsIcon";
-import Input from "../../mixins/Input";
+import Input from "./mixins/Input";
 import TextField from "./mixins/TextField";
 import FieldValidation from "./mixins/FieldValidation";
 import Helper from "../../utils/Helper";