+++ src/components/BsField/BsTextField.vue
@@ -103,7 +103,7 @@
 import { FontAwesomeIcon } from "@fortawesome/vue-fontawesome";
 import BsIcon from "../BsIcon/BsIcon";
 import BsIconToggle from "../BsIcon/BsIconToggle";
-import Input from "../../mixins/Input";
+import Input from "./mixins/Input";
 import TextField from "./mixins/TextField";
 import FieldValidation from "./mixins/FieldValidation";
 import "../../../scss/_field.scss"