+++ src/components/BsField/mixins/Input.js
@@ -1,4 +1,4 @@
-import Helper from "../utils/Helper";
+import Helper from "../../../utils/Helper";
 
 export default {
     props: {